fix(sim): widen cluster-tax simulation supply accumulators to u128 (#344)
The cluster-tax economic-simulation crate keeps its own cumulative supply
accumulators, independent of the node's ChainState. At Phase-1 scale
(~1.22e21 picocredits, well past u64::MAX ~1.84e19) these overflowed:
silently in release (overflow-checks=false) or panicking in debug. This
is the simulation-side sibling of the node fix in #333 / PR #342; it is
modeling-only with no consensus impact (cluster-tax depends only on
bth-transaction-types, not the node).

Widen the cumulative running totals that can reach supply scale to u128,
keeping per-step amounts and per-agent spot balances u64:
- MinterAgent.total_mined / total_sold (and record_minting, stats)
- SimulationState.total_rewards_emitted / total_fees_collected (and
  process_block, record_fee_burn, record_fees accumulation sites)
- MonetaryStats.total_emitted / total_fees_burned (report mirror; the
  controller's u64 totals are cast up, no narrowing introduced)
- Metrics.total_fees_collected, MetricsSummary.total_fees, and the
  snapshot_metrics total_fees param, so the reporting path carries the
  widened value with no `as u64` narrowing

The picocredit unit is unchanged (same decision as #333), so the model's
calibration is preserved and existing Gini/lottery/demurrage validations
produce identical results at sub-u64 scale.

Adds regression tests that drive each widened accumulator past u64::MAX
and assert exact accumulation with no wrap/panic (run in debug, so a
narrowed accumulator would panic):
- minter::total_mined_accumulates_past_u64_max
- state::total_rewards_emitted_accumulates_past_u64_max
- state::total_fees_collected_accumulates_past_u64_max

Closes #341

+++ cluster-tax/src/simulation/metrics.rs
@@ -35,8 +35,10 @@ pub struct Metrics {
     /// Index 0 = poorest 20%, Index 4 = richest 20%.
     pub fee_rate_by_quintile: [f64; 5],
 
-    /// Total fees collected (cumulative).
-    pub total_fees_collected: u64,
+    /// Total fees collected (cumulative). u128 to mirror the widened
+    /// `SimulationState::total_fees_collected` without narrowing at
+    /// full-supply scale (#341).
+    pub total_fees_collected: u128,
 
     /// Total number of transactions.
     pub transaction_count: u64,
@@ -206,7 +208,7 @@ impl SimulationMetrics {
 pub struct MetricsSummary {
     pub initial_gini: f64,
     pub final_gini: f64,
-    pub total_fees: u64,
+    pub total_fees: u128,
     pub total_transactions: u64,
     pub avg_fee_by_quintile: [f64; 5],
     pub wash_trade_attempts: u64,
@@ -357,7 +359,7 @@ pub fn snapshot_metrics(
     round: u64,
     agent_data: &[(AgentId, u64, u32, TagVector)], // (id, balance, fee_rate, tags)
     cluster_wealth: &ClusterWealth,
-    total_fees: u64,
+    total_fees: u128,
     transaction_count: u64,
     mixer_volume: u64,
     _fee_curve: &FeeCurve,
